+++ apps/block_scout_web/lib/block_scout_web/views/api/rpc/block_view.ex
@@ -23,6 +23,14 @@ defmodule BlockScoutWeb.API.RPC.BlockView do
     RPCView.render("show.json", data: data)
   end
 
+  def render("getblocknobytime.json", %{block_number: block_number}) do
+    data = %{
+      "blockNumber" => to_string(block_number)
+    }
+
+    RPCView.render("show.json", data: data)
+  end
+
   def render("eth_block_number.json", %{number: number, id: id}) do
     result = EthRPC.encode_quantity(number)